Also use ints for num meshes and mesh index

+++ TombLib/TombLib/LevelData/Compilers/TombEngine/Structs.cs
@@ -541,8 +541,8 @@ namespace TombLib.LevelData.Compilers.TombEngine
     public struct TombEngineMoveable
     {
         public int ObjectID;
-        public short NumMeshes;
-        public short StartingMesh;
+        public int NumMeshes;
+        public int StartingMesh;
         public int MeshTree;
         public int FrameOffset;
         public int Animation;
